Fixed 2 bugs: 1) MIME encoded filenames were previously omitted and 2) duplicate filenames with unequal hashes were poorly renamed
Issue 1: 
E-Mails can contain attachment files that have filenames which are expressed in MIME encoded-word syntax (as per RFC2047). If this is the case, the previous version of the script would not detect the file as one of the file types that are supposed to be extracted (PDF, Word, etc.). 

Fix: The filename is now converted from MIME encoded-word syntax to a human-readable filename and renamed prior to writing it to disc.


Issue 2:
When the script finds two files that have the same filename, but contain different content (i.e. the files with the same filename have different hashes), the script renames the subsequent files in the wrong way. 
For Example: three files with the same filename, but different hashes are extracted from the MBOX file - e.g.  "Invoice.pdf". The previous version of the script will then extract these PDF files and leave the filename as is for the first file ("Invoice.pdf"), but the second file will be renamed "Invoice.pdf.1" and the third file will be renamed "Invoice.pdf.2". This creates a problem, because it modifies the file extension. 

Fix: The updated script now renames the part of the filename prior to the file extension, so the second file will be renamed to "Invoice1.pdf" instead of "Invoice.pdf.1" and the third file should be renamed to "Invoice2.pdf" instead of "Invoice.pdf.2" and so on.