Fixed an error in the decode method causing the script to fail on already decoded source strings.
Issue:
If the `decode` method is being called on a string that is already decoded, it cases the script to produce an Attribute Error This happens e.g. if the `email.header.decode_header(s)` function returns a string instead of bytes for parts of the header.

Fix:
Fixed this by checking the type of `text` before calling `decode`: `isinstance(text, bytes)` checks if `text` is bytes. If it is, it is decoded. If it's not (i.e. it's already a string), the script leaves it as is.